重写指示器

+++ DanDanPlayForMac/ViewController/AutoUpdateCell.m
@@ -42,7 +42,7 @@
 
 - (IBAction)clickCheakUpdateInfoButton:(NSButton *)sender {
     [UpdateNetManager latestVersionWithCompletionHandler:^(VersionModel *model) {
-        if ([ToolsManager appVersion] < model.version.floatValue) {
+        if ([ToolsManager appVersion] < model.version) {
             NSViewController *vc = NSApp.keyWindow.contentViewController;
             [vc presentViewControllerAsModalWindow:[UpdateViewController viewControllerWithModel:model]];
         }
